requirements.txt
giref==3.3.1
autopep8==1.5.4
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
cssselect2==0.4.1
defusedxml==0.6.0
dj-rest-auth==1.1.2
Django==3.1.3
django-allauth==0.44.0
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
pathlib==1.0.1
pdf2image==1.14.0
Pillow==8.0.1
preview-generator==0.15.4
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
PyPDF2==1.26.0
python-magic==0.4.18
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
Wand==0.6.5
webencodings==0.5.1
whitenoise==5.2.0
xvfbwrapper==0.2.9
